distributed/requirements.txt
jina[http]
pydantic
pyyaml
chevron
